+++ angular-workspace/projects/ni/nimble-angular/label-provider/table/tests/nimble-label-provider-table-with-defaults.directive.spec.ts
@@ -39,6 +39,7 @@ describe('Nimble LabelProviderTable withDefaults directive', () => {
             [computeMsgId('Select all rows in group', 'Nimble table - select all rows in group')]: 'Translated select all rows in group',
             [computeMsgId('Select row', 'Nimble table - select row')]: 'Translated select row',
             [computeMsgId('Row operations', 'Nimble table - row operation column')]: 'Translated row operations',
+            [computeMsgId('Loading', 'Nimble table - row loading')]: 'Translated loading',
         });
         const fixture = TestBed.createComponent(TestHostComponent);
         const testHostComponent = fixture.componentInstance;
@@ -58,5 +59,6 @@ describe('Nimble LabelProviderTable withDefaults directive', () => {
         expect(labelProvider.groupSelectAll).toBe('Translated select all rows in group');
         expect(labelProvider.rowSelect).toBe('Translated select row');
         expect(labelProvider.rowOperationColumn).toBe('Translated row operations');
+        expect(labelProvider.rowLoading).toBe('Translated loading');
     });
 });