+++ packages/nimble-blazor/NimbleBlazor/Components/NimbleTable.razor.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text.Json;
 using Microsoft.AspNetCore.Components;
+using Microsoft.Extensions.Options;
 using Microsoft.JSInterop;
 
 namespace NimbleBlazor;
@@ -13,13 +14,12 @@ namespace NimbleBlazor;
 public partial class NimbleTable<TData> : ComponentBase
 {
     private ElementReference _table;
-    private bool _dataUpdated = false;
-    private IEnumerable<TData> _data = Enumerable.Empty<TData>();
     internal static string SetTableDataMethodName = "NimbleBlazor.Table.setData";
     internal static string GetSelectedRecordIdsMethodName = "NimbleBlazor.Table.getSelectedRecordIds";
     internal static string SetSelectedRecordIdsMethodName = "NimbleBlazor.Table.setSelectedRecordIds";
     internal static string CheckTableValidityMethodName = "NimbleBlazor.Table.checkValidity";
     internal static string GetTableValidityMethodName = "NimbleBlazor.Table.getValidity";
+    internal static string SetRecordHierarchyOptionsMethodName = "NimbleBlazor.Table.setRecordHierarchyOptions";
 
     [Inject]
     private IJSRuntime? JSRuntime { get; set; }
@@ -36,31 +36,18 @@ public partial class NimbleTable<TData> : ComponentBase
     [Parameter]
     public RenderFragment? ChildContent { get; set; }
 
-    /// <summary>
-    /// Gets or sets the data for the table.
-    /// </summary>
-    [Parameter]
-    public IEnumerable<TData> Data
-    {
-        get
-        {
-            return _data;
-        }
-        set
-        {
-            _data = value;
-            _dataUpdated = true;
-        }
-    }
-
-    /// <summary>
-    /// Gets or sets a callback that's invoked when the data changes
-    /// </summary>
-    [Parameter]
-    public EventCallback<IEnumerable<TData>> DataChanged { get; set; }
-
     [Parameter(CaptureUnmatchedValues = true)]
     public IDictionary<string, object>? AdditionalAttributes { get; set; }
+
+    /// <summary>
+    /// Sets the data in the table.
+    /// </summary>
+    /// <param name="data">The data to set in the table</param>
+    public async Task SetDataAsync(IEnumerable<TData> data)
+    {
+        var options = new JsonSerializerOptions { MaxDepth = 3 };
+        await JSRuntime!.InvokeVoidAsync(SetTableDataMethodName, _table, JsonSerializer.Serialize(data, options));
+    }
 
     /// <summary>
     /// Returns the set of selected record IDs.
@@ -77,6 +64,15 @@ public partial class NimbleTable<TData> : ComponentBase
     public async Task SetSelectedRecordIdsAsync(IEnumerable<string> recordIds)
     {
         await JSRuntime!.InvokeAsync<TableValidity>(SetSelectedRecordIdsMethodName, _table, recordIds);
+    }
+
+    /// <summary>
+    /// Sets the hierarchy options for each record in the table.
+    /// </summary>
+    /// <param name="options">The hierarchy options</param>
+    public async Task SetRecordHierarchyOptionsAsync(IEnumerable<TableSetRecordHierarchyOptions> options)
+    {
+        await JSRuntime!.InvokeVoidAsync(SetRecordHierarchyOptionsMethodName, _table, options);
     }
 
     /// <summary>
@@ -120,6 +116,12 @@ public partial class NimbleTable<TData> : ComponentBase
     public EventCallback<TableColumnConfigurationEventArgs> ColumnConfigurationChange { get; set; }
 
     /// <summary>
+    /// Gets or sets a callback that's invoked when a column's configuration is changed.
+    /// </summary>
+    [Parameter]
+    public EventCallback<TableRowExpandToggleEventArgs> RowExpandToggle { get; set; }
+
+    /// <summary>
     /// Called when 'action-menu-toggle' changes on the web component.
     /// </summary>
     /// <param name="eventArgs">The state of the action menu on the table</param>
@@ -155,15 +157,12 @@ public partial class NimbleTable<TData> : ComponentBase
         await ColumnConfigurationChange.InvokeAsync(eventArgs);
     }
 
-    /// <inheritdoc/>
-    /// <exception cref="JsonException"></exception>
-    protected override async Task OnAfterRenderAsync(bool firstRender)
+    /// <summary>
+    /// Called when the 'row-expand-toggle' event is fired on the web component.
+    /// </summary>
+    /// <param name="eventArgs">The toggle state of a table row</param>
+    protected async void HandleRowExpandToggle(TableRowExpandToggleEventArgs eventArgs)
     {
-        var options = new JsonSerializerOptions { MaxDepth = 3 };
-        if (_dataUpdated)
-        {
-            await JSRuntime!.InvokeVoidAsync(SetTableDataMethodName, _table, JsonSerializer.Serialize(Data, options));
-        }
-        _dataUpdated = false;
+        await RowExpandToggle.InvokeAsync(eventArgs);
     }
 }