+++ packages/nimble-components/src/button/tests/button.stories.ts
@@ -1,6 +1,6 @@
 import { html, when } from '@microsoft/fast-element';
 import { withActions } from '@storybook/addon-actions/decorator';
-import type { Meta, StoryObj } from '@storybook/html';
+import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
 import { ButtonAppearance, ButtonAppearanceVariant } from '../types';
 import { createUserSelectedThemeStory } from '../../utilities/tests/storybook';
 import {
@@ -24,21 +24,10 @@ interface ButtonArgs {
     endIcon: boolean;
 }
 
-const overviewText = `Per [W3C](https://www.w3.org/WAI/ARIA/apg/patterns/button/) - A button is a widget that
-enables users to trigger an action or event, such as submitting a form, opening a dialog, canceling an
-action, or performing a delete operation.
-
-If you want a button that triggers navigation to a URL, use the \`nimble-anchor-button\` instead.`;
-
 const metadata: Meta<ButtonArgs> = {
     title: 'Components/Button',
-    decorators: [withActions],
+    decorators: [withActions<HtmlRenderer>],
     parameters: {
-        docs: {
-            description: {
-                component: overviewText
-            }
-        },
         actions: {
             handles: ['click']
         }