+++ packages/nimble-components/src/icon-base/tests/icons.stories.ts
@@ -5,6 +5,7 @@ import * as nimbleIconComponentsMap from '../../icons/all-icons';
 import { IconSeverity } from '../types';
 import {
     createUserSelectedThemeStory,
+    fastParameters,
     overrideWarning
 } from '../../utilities/tests/storybook';
 import {
@@ -16,7 +17,6 @@ import { tableColumnIconTag } from '../../table-column/icon';
 import { mappingIconTag } from '../../mapping/icon';
 import { tableColumnTextTag } from '../../table-column/text';
 import { iconMetadata } from './icon-metadata';
-import { iconAddTag } from '../../icons/add';
 
 type IconName = keyof typeof nimbleIconComponentsMap;
 const data = Object.values(nimbleIconComponentsMap).map(iconClass => ({
@@ -34,14 +34,7 @@ interface IconArgs {
 const metadata: Meta<IconArgs> = {
     title: 'Components/Icons',
     parameters: {
-        docs: {
-            description: {
-                component: `Nimble icons can be slotted into other components or used independently. Each icon is available as a custom element. For example, \`<${iconAddTag}></${iconAddTag}>\``
-            },
-            source: {
-                code: null
-            }
-        }
+        ...fastParameters()
     }
 };