+++ packages/nimble-components/src/modal/index.ts
@@ -95,6 +95,19 @@ export abstract class Modal<CloseReason = void> extends FoundationElement {
         return true;
     }
 
+    /**
+     * @internal
+     */
+    public closeHandler(): void {
+        if (this.resolveShow) {
+            // If
+            // - the browser implements dialogs with the CloseWatcher API, and
+            // - the user presses ESC without first interacting with the dialog (e.g. clicking, scrolling),
+            // the cancel event is not fired, but the close event still is, and the dialog just closes.
+            this.doResolveShow(UserDismissed);
+        }
+    }
+
     // Derived classes can override this, but should not call it directly (except from the override).
     protected startOpening(): void {
         this.finishOpening();